refactor: spread out the code over multiple files
